+++ src/safe.ts
@@ -121,6 +121,8 @@ export const doubleUnderline: TemplateFormatter = safe.doubleUnderline;
 export const fg256: (n: number) => TemplateFormatter = safe.fg256;
 /** Color the foreground gray. */
 export const gray: TemplateFormatter = safe.gray;
+/** Alias of {@link gray} using the British spelling. */
+export const grey: TemplateFormatter = safe.grey;
 /** Color the foreground green. */
 export const green: TemplateFormatter = safe.green;
 /** Color the foreground bright green. */